update: Login , Logout  and loginPage authenticatoin, then only superuser can edit own message

+++ base/views.py
@@ -1,9 +1,49 @@
-from django.shortcuts import render,redirect # type: ignore
-from django.db.models import Q            # type: ignore
+from django.shortcuts import render,redirect 
+from django.contrib.auth.decorators import login_required 
+from django.db.models import Q                     
 from .models import Room,Topic
+from django.contrib.auth import authenticate,login,logout   
+from django.contrib import messages             
+from django.contrib.auth.models import User       
+from django.http import HttpResponse  
 from .forms import RoomForm
 
 # Create your views here.
+
+
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username , password=password)
+
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or password does not exist")
+
+
+    context = {}
+    return render(request, 'base/login_register.html',context)
+
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
 
 
 
@@ -31,7 +71,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request, 'base/room.html' , context)
 
-
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
 
@@ -45,10 +85,13 @@ def createRoom(request):
     return render(request,'base/room_form.html', context) 
 
 
-
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here!!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -60,8 +103,12 @@ def updateRoom(request,pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here!!")
 
     if request.method == "POST":
         room.delete()
